added getSalaryDescription in util package, change makePhoneCall in externalNavigatorImpl
